moved GoodreadsReviewAdapter to utils
